refactor: keep the package flat at the repository root
Reverts the src/ layout: the package stays at makeqr/, as it was.
The command line moves into a single makeqr/cli.py, so importing the
library no longer pulls in click.
