+++ makeqr/__main__.py
@@ -1,0 +1,4 @@
+from makeqr.cli import main
+
+if __name__ == "__main__":
+    main()
