+++ makeqr/constants.py
