+++ makeqr/makeqr.py
