+++ makeqr/models.py
