+++ makeqr/utils.py
