--- src/makeqr/__main__.py
@@ -1,10 +1,0 @@
-from makeqr.cli_app import make_app
-
-
-def main() -> None:
-    app = make_app()
-    app()
-
-
-if __name__ == "__main__":
-    main()
